refactor: 접근 제한자 추가
