+++ src/main/java/com/pet/domains/account/dto/request/AccountPasswordParam.java
@@ -1,10 +1,11 @@
 package com.pet.domains.account.dto.request;
 
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class AccountPasswordParam {
 
     private String password;
